add after/{ftplugin,plugin,syntax}
I was wrong in thinking that Vim doesn't load files from after/
directories in plugins -- it will load files from any after/ directories
in runtimepath at the appropriate time, but aside from loading at a
later time, it treats after/ directories like any other directories in
the runtimepath, meaning it expects the same directory structure. So we
can make use of an after/ directory in a plugin as long as we structure
that directory accordingly. I feel like I'm talking in circles, but I
think the contents of this commit demonstrate what I'm trying to say.